Move roborock base entity to separate module (#126135)
