+++ homeassistant/components/roborock/entity.py
