+++ homeassistant/components/roborock/sensor.py
@@ -37,7 +37,7 @@ from homeassistant.helpers.typing import StateType
 
 from . import RoborockConfigEntry
 from .coordinator import RoborockDataUpdateCoordinator, RoborockDataUpdateCoordinatorA01
-from .device import RoborockCoordinatedEntityA01, RoborockCoordinatedEntityV1
+from .entity import RoborockCoordinatedEntityA01, RoborockCoordinatedEntityV1
 
 
 @dataclass(frozen=True, kw_only=True)
